PM / OPP: Extend opp domains to support platform opp domain drivers
Many SoCs have a single regulator and clock that supply a device that
transition when a device requests a new operating point. Some SoCs
expand on this and have multiple regulators with sequencing and other
requirements for one single device. Example is TI OMAP5 and DRA7
platforms that have a second ABB (adaptive body bias) regulator that
must be sequenced specially for frequency up and down clock transitions
and Class0 AVS requirements that allow for adjusting the desired voltage
before it is set for the cpu-supply.

In order to deal with special cases like that and others that exist,
extend opp domains to allow platform OPP drivers to register themselves
with the OPP domain core and provide mechanisms to control these
regulator paths while still using the generic dev_pm_opp_set_rate call
used by the higher level generic drivers like cpufreq / devfreq.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>